# ollantaweb/docs/swagger.yaml
basePath: /
definitions:
...
  api.issueChangelogResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/postgres.ChangelogEntry'
        type: array
    type: object
  api.issueListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/postgres.IssueRow'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  api.measureTrendResponse:
    properties:
      component:
        type: string
      from:
        type: string
      metric:
        type: string
      points:
        items:
          $ref: '#/definitions/postgres.TrendPoint'
        type: array
      project:
        type: string
      to:
        type: string
    type: object
...
  postgres.ChangelogEntry:
    properties:
      created_at:
        type: string
      field:
        type: string
      id:
        type: integer
      issue_id:
        type: integer
      new_value:
        type: string
      old_value:
        type: string
      user_id:
        type: integer
    type: object
...
  postgres.IssueFacets:
    properties:
      by_directory:
        additionalProperties:
          type: integer
        type: object
      by_engine_id:
        additionalProperties:
          type: integer
        type: object
      by_file:
        additionalProperties:
          type: integer
        type: object
      by_language:
        additionalProperties:
          type: integer
        type: object
      by_lifecycle:
        additionalProperties:
          type: integer
        type: object
      by_quality:
        additionalProperties:
          type: integer
        type: object
      by_rule:
        additionalProperties:
          type: integer
        type: object
      by_security_category:
        additionalProperties:
          type: integer
        type: object
      by_severity:
        additionalProperties:
          type: integer
        type: object
      by_status:
        additionalProperties:
          type: integer
        type: object
      by_tags:
        additionalProperties:
          type: integer
        type: object
      by_type:
        additionalProperties:
          type: integer
        type: object
    type: object
  postgres.IssueRow:
    properties:
      column:
        type: integer
      component_path:
        type: string
      created_at:
        type: string
      effort_minutes:
        type: integer
      end_column:
        type: integer
      end_line:
        type: integer
      engine_id:
        type: string
      id:
        type: integer
      language:
        type: string
      line:
        type: integer
      line_hash:
        type: string
      message:
        type: string
      project_id:
        type: integer
      quality_domain:
        type: string
      resolution:
        type: string
      rule_key:
        type: string
      scan_id:
        type: integer
      secondary_locations:
        items:
          type: integer
        type: array
      severity:
        type: string
      status:
        type: string
      tags:
        items:
          type: string
        type: array
      tracking_state:
        type: string
      type:
        type: string
    type: object
...
  postgres.TrendPoint:
    properties:
      date:
        type: string
      value:
        type: number
    type: object
...
info:
  contact: {}
  description: Ollanta REST API documentation.
  title: Ollanta API
  version: "1.0"
paths:
...
  /api/v1/issues:
    get:
      description: Returns paginated issues with optional filters
      parameters:
      - description: Project ID
        in: query
        name: project_id
        type: integer
      - description: Scan ID
        in: query
        name: scan_id
        type: integer
      - description: Rule key
        in: query
        name: rule_key
        type: string
      - description: Severity
        in: query
        name: severity
        type: string
      - description: Issue type
        in: query
        name: type
        type: string
      - description: Quality domain
        in: query
        name: quality
        type: string
      - description: Status
        in: query
        name: status
        type: string
      - description: Tracking state
        in: query
        name: tracking_state
        type: string
      - description: Language
        in: query
        name: language
        type: string
      - description: Tag
        in: query
        name: tag
        type: string
      - description: Security category
        in: query
        name: security_category
        type: string
      - description: Directory
        in: query
        name: directory
        type: string
      - description: File path
        in: query
        name: file
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.issueListResponse'
      summary: List issues
      tags:
      - issues
  /api/v1/issues/{id}/changelog:
    get:
      description: Returns change history for an issue
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.issueChangelogResponse'
      summary: Issue changelog
      tags:
      - issues
  /api/v1/issues/{id}/transition:
    post:
      consumes:
      - application/json
      description: Change issue status/resolution
      parameters:
      - description: Issue ID
        in: path
        name: id
        required: true
        type: integer
      - description: Transition data
        in: body
        name: body
        required: true
        schema:
          properties:
            comment:
              type: string
            resolution:
              type: string
          type: object
      responses:
        "204":
          description: No Content
      summary: Transition issue
      tags:
      - issues
  /api/v1/issues/facets:
    get:
      description: Returns aggregated issue counts by dimension
      parameters:
      - description: Project ID
        in: query
        name: project_id
        type: integer
      - description: Scan ID
        in: query
        name: scan_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/postgres.IssueFacets'
      summary: Issue facets
      tags:
      - issues
...
  /api/v1/projects/{key}/measures/trend:
    get:
      description: Returns historical measure values for a project
      parameters:
      - description: Project key
        in: path
        name: key
        required: true
        type: string
      - description: Metric key
        in: query
        name: metric
        required: true
        type: string
      - description: From date (RFC3339)
        in: query
        name: from
        type: string
      - description: To date (RFC3339)
        in: query
        name: to
        type: string
      - description: Component path
        in: query
        name: component
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.measureTrendResponse'
      summary: Measure trend
      tags:
      - measures
...
swagger: "2.0"